fix: Test 3L_Invalid_url
