+++ tests/govtool-frontend/playwright/tests/3-drep-registration/dRepRegistration.loggedin.spec.ts
@@ -2,6 +2,7 @@ import { user01Wallet } from "@constants/staticWallets";
 import { faker } from "@faker-js/faker";
 import { test } from "@fixtures/walletExtension";
 import { setAllureEpic } from "@helpers/allure";
+import { invalid as mockInvalid } from "@mock/index";
 import DRepRegistrationPage from "@pages/dRepRegistrationPage";
 import { expect } from "@playwright/test";
 
@@ -55,48 +56,17 @@ test("3E. Should accept valid data in DRep form", async ({ page }) => {
 });
 
 test("3L. Should reject invalid data in DRep form", async ({ page }) => {
+  test.slow();
+
   const dRepRegistrationPage = new DRepRegistrationPage(page);
   await dRepRegistrationPage.goto();
 
-  function generateInvalidEmail() {
-    const choice = faker.number.int({ min: 1, max: 3 });
-
-    if (choice === 1) {
-      return faker.lorem.word() + faker.number + "@invalid.com";
-    } else if (choice == 2) {
-      return faker.lorem.word() + "@";
-    }
-    return faker.lorem.word() + "@gmail_com";
-  }
-  function generateInvalidUrl() {
-    const choice = faker.number.int({ min: 1, max: 3 });
-
-    if (choice === 1) {
-      return faker.internet.url().replace("https://", "http://");
-    } else if (choice === 2) {
-      return faker.lorem.word() + ".invalid";
-    }
-    return faker.lorem.word() + ".@com";
-  }
-  function generateInvalidName() {
-    const choice = faker.number.int({ min: 1, max: 3 });
-    if (choice === 1) {
-      // space invalid
-      return faker.lorem.word() + " " + faker.lorem.word();
-    } else if (choice === 2) {
-      // maximum 80 words invalid
-      return faker.lorem.paragraphs().replace(/\s+/g, "");
-    }
-    // empty invalid
-    return " ";
-  }
-
   for (let i = 0; i < 100; i++) {
     await dRepRegistrationPage.inValidateForm(
-      generateInvalidName(),
-      generateInvalidEmail(),
+      mockInvalid.name(),
+      mockInvalid.email(),
       faker.lorem.paragraph(40),
-      generateInvalidUrl()
+      mockInvalid.url()
     );
   }
 });
